Refactor into composable chunks
